add background entity + reset attack button
