+++ game/draw.h
@@ -1,4 +1,5 @@
 #pragma once
+#include "game.h"
 #include "game/icon.h"
 #include <game/player.h>
 
@@ -12,3 +13,4 @@ void draw_icon_line(Canvas *canvas, Vector pos, int amount, bool horizontal, con
 void spawn_icon(Level *level, const char *icon_id, float x, float y);
 void spawn_icon_line(Level *level, const char *icon_id, float x, float y, uint8_t amount, bool horizontal);
 extern char g_name[32];
+Entity *add_background(Level *level);
